+++ src/pdi_migration/validator/__init__.py
@@ -1,5 +1,17 @@
 from pdi_migration.validator.assess import assess_source
 from pdi_migration.validator.gaps import GapReport, build_gap_report
+from pdi_migration.validator.impact import ImpactAnalysis, build_impact_analysis
 from pdi_migration.validator.report import MigrationReport, build_report
+from pdi_migration.validator.score import MigrationScore, build_score
 
-__all__ = ["GapReport", "MigrationReport", "assess_source", "build_gap_report", "build_report"]
+__all__ = [
+    "GapReport",
+    "ImpactAnalysis",
+    "MigrationReport",
+    "MigrationScore",
+    "assess_source",
+    "build_gap_report",
+    "build_impact_analysis",
+    "build_report",
+    "build_score",
+]